Add data validation

+++ src/products/products.controller.ts
@@ -1,6 +1,8 @@
-import { Controller, Delete, Get, Post, Body, Param } from '@nestjs/common';
+import { Controller, Delete, Get, Post, Put, Body, Param, NotFoundException } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
+import { ParseUUIDPipe } from '@nestjs/common';
 
 @Controller('products')
 export class ProductsController {
@@ -12,12 +14,16 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  public getProductById(@Param('id') id: string) {
+  public getProductById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getProductById(id))
+      throw new NotFoundException('Product not found');
     return this.productsService.getProductById(id);
   }
 
   @Delete('/:id')
-  public deleteProductById(@Param('id') id: string) {
+  public deleteProductById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getProductById(id))
+      throw new NotFoundException('Product not found');
     this.productsService.deleteProductById(id);
     return { success: true };
   }
@@ -26,4 +32,18 @@ export class ProductsController {
   create(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
   }
+
+  @Put('/:id')
+  updateById(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+     if (!this.productsService.getProductById(id))
+       throw new NotFoundException('Product not found');
+
+    this.productsService.updateById(id, productData);
+    return { success: true };
+  }
+  
+  
 }
